backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
mysqlclient==2.2.8
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
razorpay==2.0.1
reportlab==4.4.10
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
cloudinary==1.42.2
django-cloudinary-storage==0.3.0

